+++ src/main/java/com/telepathicgrunt/blame/mixin/JigsawManagerAssemblerMixin.java
@@ -35,8 +35,8 @@ public class JigsawManagerAssemblerMixin {
     @Inject(method = "tryPlacingChildren(Lnet/minecraft/world/gen/feature/structure/AbstractVillagePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IIZ)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/util/Rotation;getShuffled(Ljava/util/Random;)Ljava/util/List;"),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private void storeCurrentPool2(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
-                                   int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece jigsawpiece) {
+    private void blame_storeCurrentPool1(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
+                                         int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece jigsawpiece) {
         if (jigsawpiece instanceof SingleJigsawPiece && ((SingleJigsawPieceAccessor) jigsawpiece).blame_getTemplateRL().left().isPresent()) {
             MissingNBTBlame.CALLING_POOL = ((SingleJigsawPieceAccessor) jigsawpiece).blame_getTemplateRL().left().get();
         }
@@ -44,8 +44,8 @@ public class JigsawManagerAssemblerMixin {
 
     @Inject(method = "tryPlacingChildren(Lnet/minecraft/world/gen/feature/structure/AbstractVillagePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IIZ)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/gen/feature/jigsaw/JigsawManager$IPieceFactory;create(Lnet/minecraft/world/gen/feature/template/TemplateManager;Lnet/minecraft/world/gen/feature/jigsaw/JigsawPiece;Lnet/minecraft/util/math/BlockPos;ILnet/minecraft/util/Rotation;Lnet/minecraft/util/math/MutableBoundingBox;)Lnet/minecraft/world/gen/feature/structure/AbstractVillagePiece;"))
-    private void storeCurrentPool2(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
-                                   int i, int i1, boolean b, CallbackInfo ci) {
+    private void blame_storeCurrentPool2(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
+                                         int i, int i1, boolean b, CallbackInfo ci) {
         MissingNBTBlame.CALLING_POOL = null;
     }
 
@@ -53,24 +53,24 @@ public class JigsawManagerAssemblerMixin {
     @Inject(method = "tryPlacingChildren(Lnet/minecraft/world/gen/feature/structure/AbstractVillagePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IIZ)V",
             at = @At(value = "INVOKE", target = "Lorg/apache/logging/log4j/Logger;warn(Ljava/lang/String;Ljava/lang/Object;)V", ordinal = 0, remap = false),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private void printMissingPoolDetails1(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
-                                          int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece structurePoolElement, BlockPos blockPos,
-                                          Rotation blockRotation, JigsawPattern.PlacementBehaviour projection, boolean bl2,
-                                          MutableObject<VoxelShape> mutableObject2, MutableBoundingBox blockBox, int i, Iterator<Template.BlockInfo> var14,
-                                          Template.BlockInfo structureBlockInfo, Direction direction,
-                                          BlockPos blockPos2, BlockPos blockPos3, int j, int k, ResourceLocation targetPoolId) {
+    private void blame_printMissingPoolDetails1(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
+                                                int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece structurePoolElement, BlockPos blockPos,
+                                                Rotation blockRotation, JigsawPattern.PlacementBehaviour projection, boolean bl2,
+                                                MutableObject<VoxelShape> mutableObject2, MutableBoundingBox blockBox, int i, Iterator<Template.BlockInfo> var14,
+                                                Template.BlockInfo structureBlockInfo, Direction direction,
+                                                BlockPos blockPos2, BlockPos blockPos3, int j, int k, ResourceLocation targetPoolId) {
         MissingTemplatePoolBlame.addEmptyPoolDetails(targetPoolId, ((SingleJigsawPieceAccessor) structurePoolElement).blame_getTemplateRL().left().orElse(null));
     }
 
     @Inject(method = "tryPlacingChildren(Lnet/minecraft/world/gen/feature/structure/AbstractVillagePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IIZ)V",
             at = @At(value = "INVOKE", target = "Lorg/apache/logging/log4j/Logger;warn(Ljava/lang/String;Ljava/lang/Object;)V", ordinal = 1),
             locals = LocalCapture.CAPTURE_FAILHARD, remap = false)
-    private void printMissingPoolDetails2(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
-                                          int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece structurePoolElement, BlockPos blockPos,
-                                          Rotation blockRotation, JigsawPattern.PlacementBehaviour projection, boolean bl2,
-                                          MutableObject<VoxelShape> mutableObject2, MutableBoundingBox blockBox, int i, Iterator<Template.BlockInfo> var14,
-                                          Template.BlockInfo structureBlockInfo, Direction direction,
-                                          BlockPos blockPos2, BlockPos blockPos3, int j, int k, ResourceLocation targetPoolId) {
+    private void blame_printMissingPoolDetails2(AbstractVillagePiece abstractVillagePiece, MutableObject<VoxelShape> voxelShapeMutableObject,
+                                                int i1, int i2, boolean b, CallbackInfo ci, JigsawPiece structurePoolElement, BlockPos blockPos,
+                                                Rotation blockRotation, JigsawPattern.PlacementBehaviour projection, boolean bl2,
+                                                MutableObject<VoxelShape> mutableObject2, MutableBoundingBox blockBox, int i, Iterator<Template.BlockInfo> var14,
+                                                Template.BlockInfo structureBlockInfo, Direction direction,
+                                                BlockPos blockPos2, BlockPos blockPos3, int j, int k, ResourceLocation targetPoolId) {
         MissingTemplatePoolBlame.addEmptyPoolDetails(targetPoolId, ((SingleJigsawPieceAccessor) structurePoolElement).blame_getTemplateRL().left().orElse(null));
     }
 }