+++ src/main/java/com/telepathicgrunt/blame/mixin/JigsawManagerMixin.java
@@ -34,12 +34,12 @@ public class JigsawManagerMixin {
     @Inject(method = "addPieces(Lnet/minecraft/util/registry/DynamicRegistries;Lnet/minecraft/world/gen/feature/structure/VillageConfig;Lnet/minecraft/world/gen/feature/jigsaw/JigsawManager$IPieceFactory;Lnet/minecraft/world/gen/ChunkGenerator;Lnet/minecraft/world/gen/feature/template/TemplateManager;Lnet/minecraft/util/math/BlockPos;Ljava/util/List;Ljava/util/Random;ZZ)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/world/gen/feature/jigsaw/JigsawPattern;getRandomTemplate(Ljava/util/Random;)Lnet/minecraft/world/gen/feature/jigsaw/JigsawPiece;"),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private static void storeCurrentPool(DynamicRegistries dynamicRegistries, VillageConfig villageConfig,
-                                         JigsawManager.IPieceFactory iPieceFactory, ChunkGenerator chunkGenerator,
-                                         TemplateManager templateManager, BlockPos blockPos,
-                                         List<? super AbstractVillagePiece> piecesList, Random random,
-                                         boolean b, boolean b1, CallbackInfo ci, MutableRegistry<JigsawPattern> mutableregistry,
-                                         Rotation rotation, JigsawPattern jigsawpattern) {
+    private static void blame_storeCurrentPool(DynamicRegistries dynamicRegistries, VillageConfig villageConfig,
+                                               JigsawManager.IPieceFactory iPieceFactory, ChunkGenerator chunkGenerator,
+                                               TemplateManager templateManager, BlockPos blockPos,
+                                               List<? super AbstractVillagePiece> piecesList, Random random,
+                                               boolean b, boolean b1, CallbackInfo ci, MutableRegistry<JigsawPattern> mutableregistry,
+                                               Rotation rotation, JigsawPattern jigsawpattern) {
         JigsawStartPoolBlame.printStartPoolCrash(jigsawpattern);
         MissingNBTBlame.CALLING_POOL = jigsawpattern.getName();
     }