+++ src/main/java/com/telepathicgrunt/blame/mixin/SingleJigsawPieceMixin.java
@@ -21,7 +21,7 @@ public class SingleJigsawPieceMixin {
 
     @Inject(method = "getTemplate(Lnet/minecraft/world/gen/feature/template/TemplateManager;)Lnet/minecraft/world/gen/feature/template/Template;",
             at = @At(value = "HEAD"))
-    private void storeCurrentPool(TemplateManager templateManagerIn, CallbackInfoReturnable<MutableBoundingBox> cir) {
+    private void blame_storeCurrentPool(TemplateManager templateManagerIn, CallbackInfoReturnable<MutableBoundingBox> cir) {
         if (MissingNBTBlame.CALLING_POOL != null & ((SingleJigsawPieceAccessor) this).blame_getTemplateRL().left().isPresent()) {
             MissingNBTBlame.storeCurrentIdentifiers(new Pair<>(MissingNBTBlame.CALLING_POOL, ((SingleJigsawPieceAccessor) this).blame_getTemplateRL().left().get()));
         }