+++ src/main/java/com/telepathicgrunt/blame/mixin/WorldEntitySpawnerMixin.java
@@ -33,20 +33,20 @@ public class WorldEntitySpawnerMixin {
     @Inject(method = "getRandomSpawnMobAt(Lnet/minecraft/world/server/ServerWorld;Lnet/minecraft/world/gen/feature/structure/StructureManager;Lnet/minecraft/world/gen/ChunkGenerator;Lnet/minecraft/entity/EntityClassification;Ljava/util/Random;Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/MobSpawnInfo$Spawners;",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/util/WeightedRandom;getRandomItem(Ljava/util/Random;Ljava/util/List;)Lnet/minecraft/util/WeightedRandom$Item;", ordinal = 0),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private static void checkIfMobSpawnWillCrash(ServerWorld serverWorld, StructureManager structureManager,
-                                                 ChunkGenerator chunkGenerator, EntityClassification entityClassification,
-                                                 Random random, BlockPos pos, CallbackInfoReturnable<MobSpawnInfo.Spawners> cir,
-                                                 Biome biome, List<MobSpawnInfo.Spawners> list) {
+    private static void blame_checkIfMobSpawnWillCrash1(ServerWorld serverWorld, StructureManager structureManager,
+                                                        ChunkGenerator chunkGenerator, EntityClassification entityClassification,
+                                                        Random random, BlockPos pos, CallbackInfoReturnable<MobSpawnInfo.Spawners> cir,
+                                                        Biome biome, List<MobSpawnInfo.Spawners> list) {
         addMobCrashDetails(serverWorld, entityClassification, pos, biome, list);
     }
 
     @Inject(method = "spawnMobsForChunkGeneration(Lnet/minecraft/world/IServerWorld;Lnet/minecraft/world/biome/Biome;IILjava/util/Random;)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/util/WeightedRandom;getRandomItem(Ljava/util/Random;Ljava/util/List;)Lnet/minecraft/util/WeightedRandom$Item;", ordinal = 0),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private static void checkIfMobSpawnWillCrash2(IServerWorld serverWorld, Biome biome,
-                                                  int chunkX, int chunkZ, Random random,
-                                                  CallbackInfo ci, MobSpawnInfo mobspawninfo,
-                                                  List<MobSpawnInfo.Spawners> list) {
+    private static void blame_checkIfMobSpawnWillCrash2(IServerWorld serverWorld, Biome biome,
+                                                        int chunkX, int chunkZ, Random random,
+                                                        CallbackInfo ci, MobSpawnInfo mobspawninfo,
+                                                        List<MobSpawnInfo.Spawners> list) {
         addMobCrashDetails(serverWorld, EntityClassification.CREATURE, chunkX, chunkZ, biome, list);
     }
 }